refactor: bid 도메인 예외 패키지명 exceptions → exception 통일
