+++ src/main/java/com/example/auction/domain/bid/service/BidCommandProcessor.java
@@ -11,7 +11,7 @@ import com.example.auction.domain.bid.dto.response.BidCachedResponse;
 import com.example.auction.domain.bid.dto.response.BidResponse;
 import com.example.auction.domain.bid.entity.Bid;
 import com.example.auction.domain.bid.enums.BidAuctionStatus;
-import com.example.auction.domain.bid.exceptions.BidErrorEnum;
+import com.example.auction.domain.bid.exception.BidErrorEnum;
 import com.example.auction.domain.bid.repository.BidRepository;
 import com.example.auction.domain.notification.dto.NotificationMessage;
 import com.example.auction.domain.notification.enums.NotificationType;
